### 2020-09-12 #2667 @katjaq
Hello @oliviaguest ! Thank you so much for being editor for our paper.
As reviewers, we would like to suggest Anibal Solon (anibalsolon), Katie Bottenhorn (62442katieb), Felix Henninger (felixhenninger). Finally, Paul McCarthy (pauldmccarthy): he's the developer of FSLeyes, however, he's not in the reviewer pool for JOSS yet...
