### 2020-09-12 6d1064f
Update paper.md
