+++ paper.md
@@ -13,7 +13,7 @@ authors:
     affiliation: "1, 2"
   - name: Roberto Toro^[Corresponding author]  
     orcid: 0000-0002-6671-858X  
-    affiliation: "2, 3"  
+    affiliation: "1, 3"  
 affiliations:
  - name: Center for Research and Interdisciplinarity, University of Paris  
    index: 1
