### 2020-09-12 #2667 @katjaq
I have a question regarding the indication of corresponding authors: Is there a way to simply have an asterisk twice, for Roberto and me to indicate corresponding authors instead of star and cross?

I would like to just have
Katja Heuer* and Roberto Toro*.

What it does though is
Katja Heuer* and Roberto Toro†

The syntax I use is
Katja Heuer^[Corresponding author] and same later for Roberto Toro^[Corresponding author]. I am sorry I cannot figure it out :(

Thank you in case you can help. :)

(Or maybe in case of 2 authors both corresponding author you don't indicate a corresponding author at all? ) 👀  Happy to follow your advice 🙃
