### 2020-09-12 #2667 @oliviaguest
I'm hoping @danielskatz can give his opinion on the above (https://github.com/openjournals/joss-reviews/issues/2667#issuecomment-691519629), as I have no idea — sorry!
