### 2020-09-13 #2670 @whedon
[REVIEW]: Reorient: A Web tool for reorienting and cropping MRI data.
**Submitting author:** @katjaq (<a href="http://orcid.org/0000-0002-7237-0196">Katja Heuer</a>)
**Repository:** <a href="https://github.com/neuroanatomy/reorient" target ="_blank">https://github.com/neuroanatomy/reorient</a>
**Version:** v1.0.0
