### 2020-09-13 #2670 @oliviaguest
Hi all! 👋 Thank you so much, @sneakers-the-rat, @pauldmccarthy for accepting to review this. Please read the instructions above. Any questions, feedback on the paper, etc., please post here. Any very code-specific questions, suggestions, etc., please use the issues in the code repo and link to them from this thread so we can all keep track of them. 🌸

For an example of how this process plays out feel free to skim previous reviews, such as: #2285 and #2348. ☺️
