### 2020-09-13 #2667 @arfon
> (Or maybe in case of 2 authors both corresponding author you don't indicate a corresponding author at all? ) 👀 Happy to follow your advice 🙃

I'm afraid our LaTeX template doesn't support denoting more than one author as a corresponding author so I'd suggest not using the footnotes like this for this case.
