+++ Pull_Request_Template.md
@@ -1,0 +1,21 @@
+<!-- Thank you for your contribution to Reorient -->
+
+<!-- Give a short title and description for your pull request: -->
+
+---
+<!-- Please run the tests. Replace each `[ ]` by `[X]` when the step is complete.-->
+- [ ] These changes fix #__ (github issue number if applicable).
+- [ ] ```npm test``` ran with full success.
+- [ ] ```npm test``` resulted in failure in _____
+
+<!-- Replace `__` with appropriate information: -->
+- [ ] I implemented tests for the changes I made OR
+- [ ] These changes do not require tests because _____
+
+<!-- Please make sure that "Allow edits from maintainers" checkbox is checked.-->
+
+<!-- Pull requests that do not address these steps are also welcome, but they will require additional verification. -->
+
+
+
+
