### 2020-09-16 421cb3b
Update paper.md
