### 2020-09-16 #2667 @katjaq
> > (Or maybe in case of 2 authors both corresponding author you don't indicate a corresponding author at all? ) 👀 Happy to follow your advice 🙃
>
> I'm afraid our LaTeX template doesn't support denoting more than one author as a corresponding author so I'd suggest not using the footnotes like this for this case.

Thank you for the quick reply, Arfon.

@oliviaguest : I deleted the corresponding author part from the paper.md file (nothing in the paper changed)--> Do I ask the whedon bot here in this thread to recompile the paper, or over in the review thread? Or do you ask for recompilation as it is already under review? Thank you :)
