+++ paper.bib
@@ -12,17 +12,17 @@
 }
 
 @article{PrimeDE, 
-    title = {An Open Resource for Non-human Primate Imaging}, 
-    volume = {100}, 
-    url = {http://dx.doi.org/10.1016/j.neuron.2018.08.039}, 
-    DOI = {10.1016/j.neuron.2018.08.039}, 
-    number = {1}, 
-    journal = {Neuron}, 
-    publisher = {Elsevier BV}, 
-    author = {Milham, Michael P. and Ai, Lei and Koo, Bonhwang and Xu, Ting and Amiez, Céline and Balezeau, Fabien and Baxter, Mark G. and Blezer, Erwin L.A. and Brochier, Thomas and Chen, Aihua and et al.},
-    year = {2018}, 
-    month = {Oct}, 
-    pages = {61-74.e2} 
+    title = {Accelerating the Evolution of Nonhuman Primate Neuroimaging},
+    volume = {105},
+    url = {http://dx.doi.org/10.1016/j.neuron.2019.12.023},
+    DOI = {10.1016/j.neuron.2019.12.023},
+    number = {4},
+    journal = {Neuron},
+    publisher = {Elsevier BV},
+    author =  {Milham, Michael and Petkov, Christopher I. and Margulies, Daniel S. and Schroeder, Charles E. and Basso, Michele A. and Belin, Pascal and Fair, Damien A. and Fox, Andrew and Kastner, Sabine and Mars, Rogier B. and et al.},
+    year = {2020},
+    month = {Feb},
+    pages = {600–603}
 }
 
 @misc{Brochier_etal2019, 
