### 2020-09-30 #2670 @sneakers-the-rat
yes give me the same - setting reminder for 10/7
