### 2020-10-03 neuroanatomy/reorient#3 @pauldmccarthy
BUG: Missing spaces/newlines in saved rotation matrix/selection box?
## Thank you for your feedback.

Hi @r03ert0 and @katjaq, this is a really nice tool! I've been playing around with it as part of openjournals/joss-reviews#2670, and have found a couple of strange behaviours - I will open one or two more issues after this one (sorry for the deluge!).

I am not 100% sure if they are bugs, or browser-specific quirks, but I thought I'd bring them to your attention just in case.

### Current behaviour

When using the _Save Rotation Matrix_ and _Save Selection Box_ options, I end up with text files which appear to be missing delimiters (probably new lines). An example rotation matrix which I get is ([also attached](https://github.com/neuroanatomy/reorient/files/5322113/reorient.txt)):

```
-0.8210228572672786 0.5441567480768857 -0.1726699202729373 35.392545939637480.42533908247618385 0.7847880104434116 0.45077094358730546 -155.19998034522737-0.38079933398443844 -0.2966499826169898 0.8757800266335992 -22.7527181620778830 0 0 1
```

Note that there are no newlines in the file - I think that the following is what should be produced:

```
-0.8210228572672786 0.5441567480768857 -0.1726699202729373 35.39254593963748
0.42533908247618385 0.7847880104434116 0.45077094358730546 -155.19998034522737
-0.38079933398443844 -0.2966499826169898 0.8757800266335992 -22.752718162077883
0 0 0 1
```

Similar behaviour occurs when using the _Save Selection Box_ option - I end up with a `selection.txt` file which looks like this ([also attached](https://github.com/neuroanatomy/reorient/files/5322112/selection.txt)):

```
-30 -96 -7930 30 67
```

But I think what is probably correct is this:

```
-30 -96 -79
30 30 67
```
### Version information (for bug reports)

* **Browser + version**: Firefox 80.0.1
* **Your OS + version**: Ubuntu 18.04
