### 2020-10-03 neuroanatomy/reorient#4 @pauldmccarthy
BUG: Interpolation bias or affine adjustment issue when reorienting LAS image
## Thank you for your feedback.

Hi again @r03ert0 and @katjaq :) I _think_ I've found a slight issue with the way that the `reorient` tool resamples an image, or constructs the voxel-to-world affine, when reorienting an image with LAS voxel storage order (the MNI152 2mm standard template which comes as part of FSL has this voxel storage order).

The `reorient` tool appears to be invariant to the voxel-storage order of loaded images (i.e. how the voxel coordinate system axes map to the LR/AP/IS anatomical coordinate system axes), and seems to always output images with a RAS voxel storage order. Overall I think this is a very good thing, as images with non LAS/RAS voxel storage orders are a common source of registration failure, so forcing all output images to have a consistent and standard storage order will help to prevent such failures.

However, there appears to be an issue in the interpolation/resampling implemented within `reorient` when working with an image which originally has LAS voxel storage order (i.e. where a left-right flip is encoded into the voxel-to-world affine matrix).  Steps to reproduce are outlined below.

### Current behaviour

1. Load an image with LAS voxel storage order into `reorient` - [here is an example](https://github.com/neuroanatomy/reorient/files/5322132/MNI152_T1_2mm_downsampled.nii.gz). It is the MNI152 2mm standard template from FSL, downsampled by a factor of ~2, to make the effect clearer.
2. Use the *Select* tool to crop the image. I cropped the image using [this selection box](https://github.com/neuroanatomy/reorient/files/5322135/selection_mni.txt) (this file has had a new line added as per #3)
3. Save the result out to disk - [here is what I get](https://github.com/neuroanatomy/reorient/files/5322137/reoriented.nii.gz)
4. View the results in an image viewer which aligns images according to their voxel-to-world affine (e.g. FSLeyes).

Here is a screencast demonstrating the result in FSLeyes (apologies for the clunkiness). I have the original MNI152 2mm template, the downsampled version, and then the re-oriented version on top:

![screencast](https://user-images.githubusercontent.com/236128/94999293-56963680-05b0-11eb-859e-d3f31512ebd2.gif)

### Version information (for bug reports)

* **Browser + version**: Firefox 80.0.1
* **Your OS + version**: Ubuntu 18.04
