### 2020-10-03 neuroanatomy/reorient#3 @r03ert0
Thank you @pauldmccarthy !
How strange. We hadn't checked with Firefox, only Chrome (and our tests run on Chromium).
Here's the `saveMatrix` function, which does have a `\n` for line breaks:

```js
function saveMatrix() {
  const {mv} = globals;
  const a = document.createElement('a');
  const m = mv.mri.MatrixVox2Mm;
  const str = m.map((o) => o.join(' ')).join('\n');
  a.href = 'data:text/csv;charset=utf-8,' + str;
  const name = prompt("Save Voxel To World Matrix (the inverse of the one displayed) As...", "reorient.mat");
  if(name !== null) {
    a.download=name;
    document.body.appendChild(a);
    a.click();
  }
}
```

We'll check in Firefox (also Safari, but maybe not Explorer :p [maybe we can add a warning])
