### 2020-10-03 neuroanatomy/reorient#3 @pauldmccarthy
I just tested on google chrome (still using Ubuntu 18.04), and indeed, the files save correctly :) So it seems to be firefox-specific. I'm not a Javascript expert, but your code looks pretty straightforward, so I can't imagine what is going on ...
