### 2020-10-03 e2e6eb9
Fix encoding of newline
