### 2020-10-03 neuroanatomy/reorient#3 @r03ert0
Fixed in commit e2e6eb9
The plain text string goes into a URL. Chrome is ok with `\n`, but Firefox wants it properly translated to `%0A` : )
(tested in Firefox 80.0.1)
