### 2020-10-08 neuroanatomy/reorient#6 @sneakers-the-rat
RangeError: invalid array length without user notification in MNI space
## Thank you for your feedback.

### Current behaviour

* If this is a bug, please explain how to reproduce the problem
    * Which tools did you use in which sequence? - loading
    * Does the error concern the online display/ functionality or the downloaded result? - display/functionality
    * What is the type and resolution of your data? (Can you provide a link to your data?) - 200um downsample in MNI space: https://datadryad.org/stash/downloads/file_stream/223919
    * Do not include sensitive information.
    * Upload the browser error message if you are familiar with web developer tools.

```
Uncaught RangeError: invalid array length
    flattenChunks https://cdn.jsdelivr.net/npm/pako@1.0.10/dist/pako.min.js:1
    onEnd https://cdn.jsdelivr.net/npm/pako@1.0.10/dist/pako.min.js:1
    push https://cdn.jsdelivr.net/npm/pako@1.0.10/dist/pako.min.js:1
    onload https://cdn.jsdelivr.net/gh/neuroanatomy/mrijs@0.0.3/mri.js:255
    pr https://cdn.jsdelivr.net/gh/neuroanatomy/mrijs@0.0.3/mri.js:251
    loadMRIFromFile https://cdn.jsdelivr.net/gh/neuroanatomy/mrijs@0.0.3/mri.js:248
    pr https://cdn.jsdelivr.net/gh/r03ert0/mriviewerjs@v0.1.4/mriviewer.js:65
    promise callback*configure/pr< https://cdn.jsdelivr.net/gh/r03ert0/mriviewerjs@v0.1.4/mriviewer.js:61
    configure https://cdn.jsdelivr.net/gh/r03ert0/mriviewerjs@v0.1.4/mriviewer.js:57
    display https://cdn.jsdelivr.net/gh/r03ert0/mriviewerjs@v0.1.4/mriviewer.js:346
    _display https://neuroanatomy.github.io/reorient/reorient.js:609
    init https://neuroanatomy.github.io/reorient/reorient.js:661
    <anonymous> https://neuroanatomy.github.io/reorient/:229
    jQuery 8
    <anonymous> https://neuroanatomy.github.io/reorient/:210

```

Upload fails without notifying user

### Expected behaviour

I don't know how common this format is, so I don't know if not supporting it is a bug, but there should be some indication of failure -- as is the page just remains on 'uploading' message.

### Version information (for bug reports)

* **Browser + version**: Firefox 81.0.1
* **Your OS + version**: 10.15.6
