### 2020-10-08 neuroanatomy/reorient#7 @sneakers-the-rat
Large files load indefinitely
## Thank you for your feedback.

### Current behaviour

Tried to test program using this 8GB scan: https://datadryad.org/stash/downloads/file_stream/223921 , let the page sit for ~40m while i was working on review, remained hung on 'uploading'

* If this is a bug, please explain how to reproduce the problem
    * Which tools did you use in which sequence? - loading
    * Does the error concern the online display/ functionality or the downloaded result?  - display/functionality
    * What is the type and resolution of your data? (Can you provide a link to your data?) - https://datadryad.org/stash/downloads/file_stream/223921
    * Do not include sensitive information.
    * Upload the browser error message if you are familiar with web developer tools. - no message in console

### Expected behaviour

would love to have a progress bar or some indication that the loading is still happening, or else an error message indicating that it failed!

### Version information (for bug reports)

* **Browser + version**: Firefox 81.0.1
* **Your OS + version**: macOS 10.15.6
