### 2020-10-08 #2670 @sneakers-the-rat
gotcha, will clean up op in  1 sec. only copying things here where i've made a comment

tbc none of these are comments that i think are worthy of holding up the manuscript, very little touches that don't get in the way of the central function of the program

	- Can't get the contribution metrics to load, though I have no reason to doubt the author's contribution & wouldn't want to punish them for github being glitchy.

	- the software is a static page hosted on github and thus doesn't require installation.
	- I was also able to open the page locally with the npm [http-server](https://www.npmjs.com/package/http-server) without issue.

Functional claims from manuscript:
- [x] data dragged onto web interface
- [x] translate and rotate by dragging in interface
- [x] selection box to crop image
- [x] save affine matrix
- [x] save selection box
- [x] save reoriented and cropped volume
- [x] existing rotation matrices loaded
- [x] append to existing rotation matrices
	- i'm not entirely sure what it means to append an affine matrix, so i'm not sure what i should be expecting when i click it, but something happens that hopefully makes sense to the authors.

Potential bugs question mark?
- Fails without notification for image in MNI space -- since there is no description in manuscript of supported/unsupported formats/filetypes/etc. I don't consider it a bug for the purposes of the review, but that would be a good thing to document & failures like this should give some indication to the user even if the format isn't supported. raised issue at https://github.com/neuroanatomy/reorient/issues/6

ux feedback/"would be nice" recommendations:
- **strongly recommend adding progress bars when loading/saving images** - the page becomes unresponsive while loading/saving and it wasn't clear to me if the page had crashed or if i just needed to keep waiting.
- an additional set of handles for translation/rotation in addition to dragging. the current handles seem to move the image in the direction normal to the plane, but it would be nice to be able to do the 'dragging' controls with handles for when eg. the rotation doesn't behave exactly as expected.
- 'reset selection box' button
- 'load new nifti' button rather than having to refresh
- would be nice to export matrix in different formats like csv, npy, etc. (unless this .mat format is universal across libraries, which i don't know)
- display metadata of file, mostly in order to give some sense of scale to the matrix/crop box but also to keep track of what file you're working on if you're working on a bunch of similar-looking images
- keep filename when saving, perhaps appending `_reoriented` rather than always defaulting to just `reoriented.nii.gz` for the same reason as ^^

	- [x] reorientation matrix and selection box can be obtained in 1m for a typical brain -- didn't time, but i think this is a reasonable statement. tested with dolphin brain: https://braincatalogue.org/Bottlenose_dolphin , 1GB human scan: https://datadryad.org/stash/downloads/file_stream/223922 , both worked, 1GB was a bit sluggish but still useable. I was unable to get an 8GB scan (https://datadryad.org/stash/downloads/file_stream/223921) to load, and without progress bar or error message I couldn't tell why - raised issue https://github.com/neuroanatomy/reorient/issues/7. Since I'm not sure if that high-res scan is a 'typical brain' I'll still count this as passing.

	- yes and love it!!
	- web tool, no installation.
	- the docs essentially function like an example as they document a typical workflow.
	- source/methods/functions don't seem to be intended for direct access by the user, so considering 'API' to be the user interface.
	- the program is relatively simple and i was able to figure out how to use it without too much guidance, i actually didn't find the documentation until i was almost finished with the review, so i think the docs should be put in the README of the repository (or another .md file if you'd rather) or else be linked to from there. I consider the core functionality reasonably documented, but there are some minor things that could do with additional documentation:
		- this may be a personal ignorance thing, but i can't really figure out what it means to append affine matrices, may be good to document
		- what the displayed matrices are, eg. i'm not sure what 'world to voxel' vs. 'voxel to world' means
		- what the format of the generated files are: are the `reorient.mat` and `selection.txt` standard file formats? what are each of the numbers in them? I get that they're an affine transformation matrix & crop box but would be good to document the format. additional details on the output image file (or even just saying that the image is unmodified except for transformation/cropping) would also be useful.
		- what the anatomical axes are: i was able to figure out the orientations of the panels, but it would be good to explicitly document what the axes are as well as what the displayed values in the top right of each panel (eg. LR: -10 in the screenshot in the README) are.
		- what file/image formats are supported

	- the mocha tests pass:
	```
	Test Reorient
    Unit tests
      eye()
        ✓ should return identity matrix
      alpha(π/2)
        ✓ should return a 90 deg X rotation matrix
      beta(π/2)
        ✓ should return a 90 deg Y rotation matrix
      gamma(π/2)
        ✓ should return a 90 deg Z rotation matrix
      matrix2str
        ✓ should return a string representation of a matrix
    End to end tests
      Load a file
        ✓ should display title
        ✓ should display "Choose..." message
        ✓ init with test nifti file (672ms)
        ✓ has the expected dimensions (107ms)
        ✓ has the expected datatype
        ✓ can rotate the image (135ms)
        ✓ can translate the image (96ms)
  	12 passing (2s)
  	```
  	- I can't get istanbul/nyc to give me code coverage, i can't access the circleci report, and locally running circleci fails because sandboxing is not configured explicitly in the `config.yml` and chromium fails to launch. not sure if there are code coverage requirements for JOSS or just the existance of CI, but can't verify coverage. tests are a little barebones (eg. tests with single hardcoded values from example image file rather than programmatically generated tests/stress tests) but they are there and seem to test what they purport to.
	- the npm package is slightly malformed, but it's not a huge problem. is there a reason why mocha is specified separately in the `.circleci/config.yml` file rather than in the `package.json` dependencies? The `http-server` should be started by the `npm test` command rather than separately if it's required for the tests to work, would recommend something like
	```json
	"scripts": {
    	"test": "http-server . & sleep 2 && mocha --exit"
  	},
  	```

	- uses the mozilla community participation guidelines <3
	- refers users to submit issues & message on twitter which seems appropriate.
	- lack of API-level documentation makes it sorta unclear to me how i would go about contributing, but I think that's fine for this project -- the source is reasonably clear and names are clear enough that I think I could find what  was looking for.

	- yes and i still love it!
	- comparison made to automated atlas registration -> need for tool. can't tell if there are other tools like this but doesn't matter for this project.
