### 2020-10-08 #2670 @oliviaguest
@pauldmccarthy was this just one you missed out on accidentally?

>  State of the field: Do the authors describe how this software compares to other commonly-used packages?
