### 2020-10-08 #2670 @pauldmccarthy
@oliviaguest, No, I wanted to re-read the article/docs before signing off on everything - planning to finish it off tonight!
