### 2020-10-08 neuroanatomy/reorient#8 @pauldmccarthy
ENH:Uncompressed .nii files not supported?
## Thank you for your feedback.

It seems that only `.nii.gz` files can be loaded into reorient, and not uncompressed `.nii` files.

This is a admittedly a slightly unfair criticism, as the load page explicitly says *Choose a `.nii.gz` file..*, but the first file I tried to load into the tool was a `.nii` file, and I found it strange that nothing happened.

It is also true that compresesd `.nii.gz` files are more common than `.nii` files, so feel free to ignore and close this issue if you don't think it is a big deal :)

### Current behaviour

* If this is a bug, please explain how to reproduce the problem
    * Which tools did you use in which sequence?
    * Does the error concern the online display/ functionality or the downloaded result?
    * What is the type and resolution of your data? (Can you provide a link to your data?)
    * Do not include sensitive information.
    * Upload the browser error message if you are familiar with web developer tools.

### Version information (for bug reports)

* **Browser + version**: Firefox 81.0 / Google Chrome 85.0.4183.121
* **Your OS + version**: Ubuntu 18.04
