### 2020-10-08 neuroanatomy/reorient#9 @pauldmccarthy
DOC: Instructions on how to load a new file
## Thank you for your feedback.

From what I could gather, once you have reoriented one image, and wish to reorient a new image, you need to refresh the web page in order to reload the "starting" page, and the ability to load a new NIFTI image.

I don't have a problem with this workflow, but I couldn't find anything in the documentation, and I thought it might be worth mentioning somewhere. Or perhaps adding a link/button somewhere in the UI to "Load new image"

I may have missed something here, so please feel free to disregard this issue :)

### Current behaviour

* If this is a bug, please explain how to reproduce the problem
    * Which tools did you use in which sequence?
    * Does the error concern the online display/ functionality or the downloaded result?
    * What is the type and resolution of your data? (Can you provide a link to your data?)
    * Do not include sensitive information.
    * Upload the browser error message if you are familiar with web developer tools.

### Version information (for bug reports)

* **Browser + version**:
* **Your OS + version**:
