### 2020-10-08 #2670 @pauldmccarthy
Apart from a few minor issues and suggestions (listed below), I think this is a really neat piece of work!

There is just one point made in the manuscript which I would like clarification on:

> The affine matrix and selection box can be used later within a scripted workflow

I couldn't find any further mention of this point in the manuscript, nor in the source repository or app help. I presume the intent is that other tools (e.g. `fslroi`, `flirt`, etc - highlighting my bias towards FSL there) would be used for using the outputs in a scripted workflow?

I'm not necessarily requesting any changes to the manuscript or tool documentation, but it was just something which stood out to me, and which I wasn't sure what to make of, so was hoping for this point to be clarified.

Thanks!

 * neuroanatomy/reorient#3
 * neuroanatomy/reorient#4
 * neuroanatomy/reorient#8
 * neuroanatomy/reorient#9
