### 2020-10-08 #2670 @pauldmccarthy
[addendum, after looking at the extremely comprehensive review by @sneakers-the-rat] indeed, it may be worth adding an overview somewhere in the help documentation (I don't think this belongs in the manuscript) about how to use the selection box/affine files "off-line", i.e. outside of the reorient web interface, in an automated pipeline.

Unfortunately there are substantial differences across the different neuroimaging packages in the ways that files/information such as these are applied. I suppose the assumption is that if a user wished to use these outputs in their own workflow, they would know how to apply them using their tools of choice, which is fair enough. But I feel that, given that it is mentioned in the manuscript, it probably should be mentioned, if only briefly, somewhere in the documentation :)
