### 2020-10-19 872d055
fix crop box display
