### 2020-10-19 f5f8c9b
rotate and crop a nifti file using the matrix and selection from reorient
