+++ reorient.py
@@ -35,13 +35,15 @@ def trilinear(x, y, z, data, dim):
 
   return Vxyz
 
-def value_at_absolute_coordinate(a, mm2vox, data, dim):
-  v = mm2vox.dot(a)
-  [x, y, z] = [int(v[0]), int(v[1]), int(v[2])]
-  if x<0 or x>=dim[0] or y<0 or y>=dim[1] or z<0 or z>=dim[2]:
+def value_at_absolute_coordinate(absolute_coordinate, mm2vox, data, dim):
+  v = mm2vox.dot(absolute_coordinate)
+  # [x, y, z] = [int(v[0]), int(v[1]), int(v[2])]
+  # if x<0 or x>=dim[0] or y<0 or y>=dim[1] or z<0 or z>=dim[2]:
+  #   return 0
+  [x, y, z] = [v[0], v[1], v[2]]
+  if x<0 or x>=dim[0]-1 or y<0 or y>=dim[1]-1 or z<0 or z>=dim[2]-1:
     return 0
-
-  return trilinear(v[0], v[1], v[2], data, dim)
+  return trilinear(x, y, z, data, dim)
 
 def world_pixdim(nii):
   v2w = nii.affine[:3,:3]
@@ -74,7 +76,7 @@ def reorient(img, rotation, selection):
     int(np.ceil(selection[1,2] - selection[0,2]))
   ]
   mm2vox = np.linalg.inv(rotation)
-  orig_data = img.get_data()
+  orig_data = img.get_fdata()
   orig_dim = img.shape
 
   data = np.zeros(list(dim), dtype = np.float32)
@@ -101,6 +103,22 @@ def reorient(img, rotation, selection):
   return result
 
 def main(argv):
+
+  if len(argv) != 5:
+    print('''
+    reorient.py
+
+    Usage:
+      python reorient.py input rotation_file selection_file output
+
+    Where:
+      input: path to source nifti file.
+      rotation_file: path to rotation matrix generated by the reorient app.
+      selection_file: path to selection box generated by the reorient app.
+      output: path to resulting, reoriented and cropped, nifti file.
+    ''')
+    return
+
   _,input_file, rotation_file, selection_file, output_file = argv
 
   img = nib.load(input_file)
