### 2020-10-19 neuroanatomy/reorient#4 @r03ert0
I think it's fixed now : )

There were several things. We created a series of very small nifti files, with different matrices, voxel shape, etc. I'll post some examples.

The display of the crop-box was not accurate to the pixel, the image display was using nearest neighbour interpolation only, and the size of the image was also not pixel-accurate.

Here's the result on the same downsampled image:

![out](https://user-images.githubusercontent.com/2310732/96442922-28e5fa00-120c-11eb-9543-a230d0ba1588.gif)
