### 2020-10-19 neuroanatomy/reorient#4 @r03ert0
Another thing that seemed to contribute to the bias was the volume translation. Volumes resampled using reorient had by default a translation that was an integer number of voxel dimensions. In FSLEyes, moving to the (0,0,0) coordinate, for example, will put the axis cross in the middle of the voxel. The source downsampled image (MNI152...) had a different displacement. Because voxels were so big, this contributed to the "shift" effect. Here's a zoom on the previous GIF showing this:
![voxel_shift](https://user-images.githubusercontent.com/2310732/96443655-3c459500-120d-11eb-9e81-11034a2a355b.gif)

Additionally, the GIF shows the isotropic resampling of reorient compared with the original anisotropic voxels.
