### 2020-10-19 neuroanatomy/reorient#4 @r03ert0
If the voxels of the source MNI152 volume are made to be isotropic and the translation rounded to voxel size in world space (just by changing the affine matrix), the position of the voxels and their values match perfectly, showing that the code works correctly now :)
![perfect](https://user-images.githubusercontent.com/2310732/96444391-7b281a80-120e-11eb-9ef6-81d7c72eb7e5.gif)
(I used the "copper" colourmap, otherwise, there's no difference)
