### 2020-10-19 neuroanatomy/reorient#4 @r03ert0
One more test: a small volume with 2*3*4 voxels is associated to 6 different affine matrices, representing different encodings for the voxels. For each volume the image shows that the rendering is the same as in FSLEyes (L/R radiological in FSLEyes, but neurological in Reorient). The corresponding affine matrices are displayed in each subplot.

![orientation 001](https://user-images.githubusercontent.com/2310732/96468107-fb10ad80-122b-11eb-98ce-f492b1c88e18.jpeg)
