### 2020-10-19 neuroanatomy/reorient#4 @r03ert0
This image shows that the crop result is well aligned with the original image for all 6 affine matrices. Each time the crop-box is the same. In the cases where the crop-box includes voxels outside the original nifti volume, the space is filled with 0. The cropped result is displayed using the copper colourmap (otherwise, the grey levels are exactly the same).

![orientation 002](https://user-images.githubusercontent.com/2310732/96468430-63f82580-122c-11eb-84ce-2749e064a54f.jpeg)
