### 2020-10-19 neuroanatomy/reorient#4 @r03ert0
We have more test images, in particular, testing the interpolation with anisotropic voxels, but the results are similar to those from the comment 4h ago. I'll close the issue : )
Thank you @pauldmccarthy for pointing it out!
