### 2020-10-19 a87db8a
smaller default cropbox
