### 2020-10-19 neuroanatomy/reorient#8 @r03ert0
now uploading .nii files is supported as well : )
