### 2020-10-19 neuroanatomy/reorient#7 @r03ert0
Thank you for the issue report @sneakers-the-rat!
We improved the feedback given when loading a file. We now display a message indicating that an uploading is going on. In the case of the file you were trying, it failed because it's extremely large for MRI standards. In case of error, we now display a message indicating to either try again or to report a bug. We also check for file size, and provide an indication in case the file is >200MB (which is rarely the case, most brain MRI files are <100MB).
