### 2020-10-20 #2670 @r03ert0
hello! quick update. We've pushed several commits to address the issues left by @sneakers-the-rat and @pauldmccarthy. There's in particular a python script (`reorient.py`) which can take the original MRI file, plus the affine matrix and the selection box produced by the Web app, and reproduce the result of "Save Nifti" from the interface (so that there's no need for the Web app for just re-running the process automatically).
We still want to update the doc to take into account some of your comments, and then will draft a detailed answer :)
