### 2020-10-27 d9e5fd7
add "Reset Selection" button
