### 2020-10-28 c6f8ec9
added documentation for all new features and more explanation to the doc
