### 2020-10-28 #2670 @katjaq
Hello :)

A huge thank you to the reviewers for their amazing detailed reviews. We addressed all of the points mentioned, implemented the suggested add ons, and included more info in the doc as well as a pointer to the doc in the Readme. Please find the detailed  answers to each issue below.

Q: "append to existing rotation matrices i'm not entirely sure what it means to append an affine matrix, so i'm not sure what i should be expecting when i click it, but something happens that hopefully makes sense to the authors."
A: Each MRI volume has its own affine matrix. The options are to change original_matrix into loaded_matrix, or original_matrix into original_matrix*loaded_matrix. This information was added to the doc.

Q: "Fails without notification for image in MNI space -- since there is no description in manuscript of supported/unsupported formats/filetypes/etc. I don't consider it a bug for the purposes of the review, but that would be a good thing to document & failures like this should give some indication to the user even if the format isn't supported. raised issue at neuroanatomy/reorient#6"
A: Accepted file formats are now checked, and an alert displayed in case of incorrect format. Additionally, why added the possibility of reading uncompressed nii files in addition to nii.gz (Commit e2e6eb9)

Q: "strongly recommend adding progress bars when loading/saving images - the page becomes unresponsive while loading/saving and it wasn't clear to me if the page had crashed or if i just needed to keep waiting."
A: We added a message to show uploading, and better error message in case of failure (commit 7ad208f)

Q: "an additional set of handles for translation/rotation in addition to dragging. the current handles seem to move the image in the direction normal to the plane, but it would be nice to be able to do the 'dragging' controls with handles for when eg. the rotation doesn't behave exactly as expected."
A: Translation handles would be a good idea, but we haven't found a way of making a good controller for that. The traditional way of doing translation is by clicking on the center of the selection box and dragging. However, in our experience it happened that we could translate instead of move a single side by mistake, which then requires to re-adjust the whole selection. Our current approach may take a bit longer, but it's always safer... Concerning Rotation handles, our focus is currently to align the MRI to the orthogonal anterior-posterior/left-right/superior-inferior axes. A general tool for reorienting with arbitrary directions could be useful in some other contexts, however.

Q: "'reset selection box' button"
A: added in commit d9e5fd7

Q: 'load new nifti' button rather than having to refresh"
A: We discussed this at length. It would be very simple to add, but we actually prefer the reload, and having one button less in the interface. However, we do explicitly state in the documentation now that a "reload" is requited for working with a new MRI. (commit c6f8ec9)

Q: "would be nice to export matrix in different formats like csv, npy, etc. (unless this .mat format is universal across libraries, which i don't know)"
A: The .mat format is widely used in neuroimaging. It is actually just a text file with four rows of 4 numbers separated by spaces, and can be easily read using python or other languages. We included a python script for reproducing a reorientation offline (reorient.py, commit 6083a41). This info is added now to the doc as well.

Q: "display metadata of file, mostly in order to give some sense of scale to the matrix/crop box but also to keep track of what file you're working on if you're working on a bunch of similar-looking images"
A: We now display the file name on top of the MRI viewers.

Q: "keep filename when saving, perhaps appending _reoriented rather than always defaulting to just reoriented.nii.gz for the same reason as ^^"
A: This could be a feature, or a bug... In our case, we keep subject ID only in the directory name, and all files inside that directory have standard names (t2.nii.gz, dwi.nii.gz, etc). We like it because scripts look cleaner. Adding the filename to the reorient.mat or selection.txt files would force us to manually remove the file name every time... It's true that some people may have the opposite approach, and will have to manually add the names... we took a selfish decision...

Q: "reorientation matrix and selection box can be obtained in 1m for a typical brain"
A: Typical MRI file sizes go from ~10 MB to about 100 MB. In animal MRI sometimes file sizes are larger, and we do have some files which are 300 MB. We haven't tried with larger files, and it is a nice surprise to know that the browser didn't break for 1GB . In case of a loading failure, we now check if the file is too large, and provide that information as feedback to the user (commit 7ad208f).

Q: " i think the docs should be put in the README"
A: We added a link to the documentation to the Readme file. Users do not need to go to the github repo to use the tool, so they may never see the README file there. The "?" tab in the tool links to the documentation from the tool :)

Q: "what the displayed matrices are, eg. i'm not sure what 'world to voxel' vs. 'voxel to world' means"
A: This is the standard way of naming in neuroimaging a matrix that would convert voxel indices (integer numbers starting from 0) into millimeters : that's the voxel_to_world. The world_to_voxel matrix does the inverse: it takes a coordinate in millimeters from real (world) space into voxel indices. Because world coordinates can be anything, they may not map inside the voxel volume, in which case the voxel index coordinates would be either negative, or larger than the volume dimensions (in voxels). This information is now also included in the doc.

Q: "what the format of the generated files are: are the reorient.mat and selection.txt standard file formats?"
A: The .mat format is standard (although it is just a text file). The selection.txt format is not. The first row in the selection format represents the coordinates of the "minimum" corner, the second row the coordinates of the "maximum" corner. We now provide a python script that can take the source MRI, the reorient.mat file, and the selection.txt file to produce the reoriented volume (commit 6083a41). This info has also been added to the doc.

Q: "what the anatomical axes are:"
A: LR is left/right, AP is anterior/posterior, SI is superior/inferior.

Q: "is there a reason why mocha is specified separately"
A: Not really... but we added now the "test" call to the npm package as suggested.

Q: "lack of API-level documentation"
A: We used jsdoc-style comments which could allow us to generate an API documentation.

Q: "The affine matrix and selection box can be used later within a scripted workflow"
A: reorient.py script added (commit 6083a41). We also added an offline workflow description to the doc.

Q: it may be worth adding an overview somewhere in the help documentation (I don't think this belongs in the manuscript) about how to use the selection box/affine files "off-line", i.e. outside of the reorient web interface, in an automated pipeline.
A: We added a "Reproduce the workflow offline" entry in the doc.

Finally, we addressed some other issues that had been directly added in reorient's repo. In particular, @pauldmccarthy had found a bug in our data display that was evident when the volumes were very small (-> voxels very big). We fixed the issue, and tested the interface extensively: https://github.com/neuroanatomy/reorient/issues/4

Thank you very much again for reviewing our tool!
Katja & Roberto
