### 2020-10-31 #2670 @oliviaguest
@katjaq @r03ert0 Has neuroanatomy/reorient/issues/9 been addressed? Can you close if so.

Also can you deposit code on zenodo (or similar) and post the DOI here, please? ☺️
