### 2020-10-31 neuroanatomy/reorient#9 @r03ert0
hello! Yes, reload is our preferred way of loading a new MRI. Instructions were added to the doc.
