### 2020-10-31 #2670 @katjaq
Hello,
issue #9 has bee addressed and is now closed. Thanks.

The code is deposited on Zenodo under DOI 010.5281/zenodo.4165153

Thank you!
