### 2020-10-31 #2670 @oliviaguest
@katjaq @r03ert0 great! can you also change the title of the zenodo to the same as the JOSS paper, please? ☺️

Oh, and is the version number at the top correct? If not, what's the current version?
