### 2020-11-01 #2670 @r03ert0
Done! I updated the title to match that of the paper.
The DOI which appears in the Zenodo badge in reorient's repo is automatically set to the one of the latest version (10.5281/zenodo.4165154).
The DOI Katja sent is the umbrella DOI for all versions (10.5281/zenodo.4165153). It's better to use that one, no?
